+++ src/renderer/src/App.tsx
@@ -23,6 +23,7 @@ import { EvidencePage } from './pages/Evidence'
 import { CrossRef } from './pages/CrossRef'
 import { CoLocate } from './pages/CoLocate'
 import { FbId } from './pages/FbId'
+import { Wireless } from './pages/Wireless'
 import { Enumerate } from './pages/Enumerate'
 import { Timeline } from './pages/Timeline'
 import { ExamPrep } from './pages/ExamPrep'
@@ -80,6 +81,7 @@ export default function App(): JSX.Element {
               <Route path="/cross-ref" element={<CrossRef />} />
               <Route path="/co-locate" element={<CoLocate />} />
               <Route path="/fb-id" element={<FbId />} />
+              <Route path="/wireless" element={<Wireless />} />
               <Route path="/enumerate" element={<Enumerate />} />
               <Route path="/timeline" element={<Timeline />} />
               <Route path="/exam-prep" element={<ExamPrep />} />